+++ CHANGELIST.txt
@@ -50,8 +50,10 @@ To Do
 
 1.0.17
 ------
+---> Add first() to IndexedDataReference.
+
 Http Automation
-    - Http Facade class added.
+    - Http Facade class added. Sessions are created using this class. Also supports quick sender methods with temp session creation.
     - Support for basic and digest authentication.
     - The session module is split into multiple modules for better maintenance as the code became large for single file.
     - Added support for HEAD, PATCH, OPTIONS methods.
@@ -71,6 +73,10 @@ Http Automation
 Configuration
     - Added ArjunaOption.DATA_FILE_DIR to contain files that are supposed to be used as file data.
     - create-project command creates project/data/file directory with a placeholder.
+Data References
+    - Moved IndexedDataReference and ContextualDataReference classes to TPI.
+NodeLocator
+    - Now available as Xml.node_locator
 
 1.0.16
 ------
