+++ docs/source/arjuna.tpi.data.reference.rst
@@ -1,0 +1,7 @@
+arjuna.tpi.data.reference module
+================================
+
+.. automodule:: arjuna.tpi.data.reference
+   :members:
+   :undoc-members:
+   :show-inheritance:
